requirements.txt
# Core Dependencies
eel==0.16.0
pyttsx3==2.90
SpeechRecognition==3.10.0
opencv-python==4.8.1.78
mediapipe==0.10.8
cvzone==1.6.1
pyautogui==0.9.54
pynput==1.7.6
pywhatkit==5.4
keyboard==0.13.5
playsound==1.3.0
dateparser==1.2.0
plyer==2.1.0
psutil==5.9.6

# Web Framework
Flask==3.0.0

# AI & API
hugchat==0.4.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0

# Audio (optional, for wake word)
pyaudio==0.2.14
pvporcupine==3.0.2

# Additional utilities
numpy==1.26.2
